% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dia2cat.R
\encoding{UTF-8}
\name{dia2cat}
\alias{dia2cat}
\title{Title diamètre vers cat\u00e9gories de diamètre}
\usage{
dia2cat(diam)
}
\arguments{
\item{diam}{vecter de diametre}
}
\value{
vecteurs des cat\u00e9gories
}
\description{
Title diamètre vers cat\u00e9gories de diamètre
}
