% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dia2cat5cm.R
\name{dia2cat5cm}
\alias{dia2cat5cm}
\title{Title diamètre vers categories de diamètre de 5 en 5}
\usage{
dia2cat5cm(diam)
}
\arguments{
\item{diam}{vecteur de diametre}
}
\value{
tableau de la surface terrière selon categories de diamètre 5cm en 5cm
}
\description{
Title diamètre vers categories de diamètre de 5 en 5
}
