% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keclaircie.R
\name{keclaircie}
\alias{keclaircie}
\title{Nature d eclaircie}
\usage{
keclaircie(diam, mart, tarif)
}
\arguments{
\item{diam}{vecteur des diametre des arbres du peuplement en metre}

\item{mart}{vecteur factoriel des arbres marteles}

\item{tarif}{numéro dutarif algan}
}
\value{
coefficient K de la nature de l eclaircie
K = ve/vav où ve est le volume de l'arbre moyen enlevé à l'éclaircie et vav est le volume de l'arbre moyen avant éclaircie.

   Si K < 0,7 l'éclaircie est qualifiée de « par le bas »
   si K > 1, de « par le haut », et
   si 0,7 < K < 1, de « mixte »
}
\description{
Nature d eclaircie
}
