% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perrotte.R
\name{perrotte}
\alias{perrotte}
\title{formmule de perotte}
\usage{
perrotte(diam, tarif)
}
\arguments{
\item{diam}{vecteur des diametres des arbres en mètre}

\item{tarif}{numéro dutarif algan}
}
\value{
vecteurs volume des arbres en m^3
}
\description{
formmule de perotte
}
