% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dia2cat.R
\encoding{UTF-8}
\name{dia2cat}
\alias{dia2cat}
\title{Title diametre vers categorie de diametre}
\usage{
dia2cat(diam)
}
\arguments{
\item{diam}{vecteur de diametre}
}
\value{
vecteurs des categories
}
\description{
Title diametre vers categorie de diametre
}
