% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_ponderee.R
\name{dist.pond}
\alias{dist.pond}
\title{Distance moyenne ponderee}
\usage{
dist.pond(geometry, diam, tarif)
}
\arguments{
\item{diam}{vecteur des diametres des arbres en mètre}

\item{tarif}{numéro dutarif algan}

\item{geom}{Geometries des arbres = longitude et latitudes}
}
\value{
vecteurs distance moyenne des arbres ponderee par leur volume
}
\description{
Distance moyenne ponderee
}
