% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitement_essence.R
\name{traitement_essence}
\alias{traitement_essence}
\title{traitement essence}
\usage{
traitement_essence(tab, duree, surface)
}
\arguments{
\item{tab}{fichier import\u00e9}

\item{duree}{duree de rotation}

\item{surface}{surface de la parcelle}
}
\value{
table des accroissement et des passage à la futaie par essence
}
\description{
traitement essence
}
